services/detection-plane/pyproject.toml
[project]
name = "aap-detection-plane"
version = "0.1.0"
description = "AAP Detection Plane — alert classification, deduplication, and state management"
requires-python = ">=3.10"
dependencies = [
    "azure-cosmos>=4.7.0",
    "azure-identity>=1.17.0",
    "azure-mgmt-alertsmanagement>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"
